+++ kolibri/plugins/learn/assets/src/views/explore-page/index.vue
@@ -13,24 +13,26 @@
       {{ topic.description }}
     </p>
 
-    <span class="visuallyhidden" v-if="subtopics.length">{{ $tr('navigate') }}</span>
+    <ui-select
+      :label="$tr('display')"
+      :options="filterOptions"
+      v-model="selectedFilter"
+      class="filter"
+    />
 
-    <card-list v-if="subtopics.length">
-      <topic-list-item
-        v-for="topic in subtopics"
-        :title="topic.title"
-        :link="genTopicLink(topic.id)"/>
-    </card-list>
+    <span class="visuallyhidden" v-if="subtopics.length">{{ $tr('navigate') }}</span>
 
     <card-grid v-if="contents.length">
       <content-grid-item
         v-for="content in contents"
+        v-show="selectedFilter.value === 'all' || selectedFilter.value === content.kind"
+        :key="content.id"
         class="card"
         :title="content.title"
         :thumbnail="content.thumbnail"
         :kind="content.kind"
         :progress="content.progress"
-        :link="genContentLink(content.id)"/>
+        :link="genLink(content)"/>
     </card-grid>
 
   </div>
@@ -42,12 +44,23 @@
 
   const getCurrentChannelObject = require('kolibri.coreVue.vuex.getters').getCurrentChannelObject;
   const PageNames = require('../../constants').PageNames;
+  const ContentNodeKinds = require('kolibri.coreVue.vuex.constants').ContentNodeKinds;
+  const some = require('lodash/some');
+  const forEach = require('lodash/forEach');
 
   module.exports = {
     $trNameSpace: 'learnExplore',
     $trs: {
       explore: 'Topics',
       navigate: 'Navigate content using headings',
+      all: 'All content',
+      topics: 'Topics',
+      exercises: 'Exercises',
+      videos: 'Videos',
+      audio: 'Audio',
+      documents: 'Documents',
+      html5: 'HTML5 Apps',
+      display: 'Display',
     },
     components: {
       'page-header': require('../page-header'),
@@ -55,30 +68,59 @@
       'content-grid-item': require('../content-grid-item'),
       'card-grid': require('../card-grid'),
       'card-list': require('../card-list'),
+      'ui-select': require('keen-ui/src/UiSelect'),
     },
+    data: () => ({
+      selectedFilter: '',
+    }),
     computed: {
+      filterOptions() {
+        const options = [{ label: this.$tr('all'), value: 'all' }];
+        const kindLabelsMap = {
+          [ContentNodeKinds.TOPIC]: this.$tr('topics'),
+          [ContentNodeKinds.EXERCISE]: this.$tr('exercises'),
+          [ContentNodeKinds.VIDEO]: this.$tr('videos'),
+          [ContentNodeKinds.AUDIO]: this.$tr('audio'),
+          [ContentNodeKinds.DOCUMENT]: this.$tr('documents'),
+          [ContentNodeKinds.HTML5]: this.$tr('html5'),
+        };
+        forEach(kindLabelsMap, (value, key) => {
+          if (this.contentsContain(key)) {
+            options.push({ label: value, value: key });
+          }
+        });
+        return options;
+      },
       title() {
         return this.isRoot ? this.$tr('explore') : this.topic.title;
       },
+      subtopics() {
+        return this.contents.filter(content => content.kind === ContentNodeKinds.TOPIC);
+      },
     },
     methods: {
-      genTopicLink(id) {
+      contentsContain(kind) {
+        return some(this.contents, content => content.kind === kind);
+      },
+      genLink(node) {
+        if (node.kind !== ContentNodeKinds.TOPIC) {
+          return {
+            name: PageNames.EXPLORE_CONTENT,
+            params: { channel_id: this.channelId, id: node.id },
+          };
+        }
         return {
           name: PageNames.EXPLORE_TOPIC,
-          params: { channel_id: this.channelId, id },
+          params: { channel_id: this.channelId, id: node.id },
         };
       },
-      genContentLink(id) {
-        return {
-          name: PageNames.EXPLORE_CONTENT,
-          params: { channel_id: this.channelId, id },
-        };
-      },
+    },
+    mounted() {
+      this.selectedFilter = this.filterOptions[0];
     },
     vuex: {
       getters: {
         topic: state => state.pageState.topic,
-        subtopics: state => state.pageState.subtopics,
         contents: state => state.pageState.contents,
         isRoot: (state) => state.pageState.topic.id === getCurrentChannelObject(state).root_id,
         channelId: (state) => getCurrentChannelObject(state).id,
@@ -95,5 +137,9 @@
     margin-top: 1em
     margin-bottom: 1em
     line-height: 1.5em
+
+  .filter
+    width: 200px
+    margin-top: 2em
 
 </style>