+++ CHANGELOG.rst
@@ -5,6 +5,16 @@ Release Notes
 
 Changes are ordered reverse-chronologically.
 
+0.5
+---
+
+ - Update all user logging related timestamps to a custom datetime field that includes timezone info
+ - Added daemon mode (system service) to run ``kolibri start`` in background (default!) #1548
+ - Implemented ``kolibri stop`` and ``kolibri status`` #1548
+ - Newly imported channels are given a 'last_updated' timestamp
+ - Add progress annotation for topics, lazily loaded to increase page load performance
+ - Add API endpoint for getting number and total size of files in a channel
+
 
 0.4
 ---